+++ .github/workflows/pythonapp.yml
@@ -32,5 +32,6 @@ jobs:
         flake8 . --count --exit-zero --statistics
     - name: Test with pytest
       run: |
+        pip install .
         pip install pytest
         pytest
